move outlier detector
