+++ source/Sailfish/Analysis/OutlierAnalysis.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Sailfish.Analysis;
+
+public record OutlierAnalysis(
+    double[] DataWithOutliersRemoved,
+    IEnumerable<double> LowerOutliers,
+    IEnumerable<double> UpperOutliers,
+    int TotalNumOutliers);
